backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-multipart==0.0.20
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4
pydantic-settings==2.6.1
psycopg2-binary==2.9.10
google-generativeai==0.8.3
pypdfium2==4.30.1
Pillow==11.0.0
requests==2.32.3
greenlet==3.1.1
boto3==1.35.72
gunicorn==23.0.0

